+++ CHANGELOG.md
@@ -21,6 +21,8 @@ And please only add new entries to the top of this list, right below the `# Unre
 - On X11, add mappings for numpad comma, numpad enter, numlock and pause.
 - On macOS, fix Pinyin IME input by reverting a change that intended to improve IME.
 - On Windows, fix a crash with transparent windows on Windows 11.
+- Add `set_clipboard_content` and `request_clipboard_content` API on `Window`
+- Add `set_primary_clipboard_content` and `request_primary_clipboard_content` API on `WindowExtUnix`
 
 # 0.26.0 (2021-12-01)